WIP Refactor API
