+++ test/histogram_matching.jl
@@ -9,9 +9,9 @@
     # omit N0f8 and N0f16 from this particular test.
     for T in (Gray{Float32}, Gray{Float64})
         img = T.(testimage("mandril_gray"))
-        imgo = T.(adjust_histogram(GammaCorrection(),img,0.5))
+        imgo = T.(adjust_histogram(img, GammaCorrection(gamma = 0.5)))
 
-        result = adjust_histogram(Matching(),img,imgo,256)
+        result = adjust_histogram(img, Matching(targetimg = imgo, nbins = 256))
 
         edges_target, counts_target = build_histogram(imgo, 256, minval = 0, maxval = 1)
         edges_result, counts_result = build_histogram(result, 256, minval = 0, maxval = 1)
@@ -20,7 +20,7 @@
         # Check that we get the same result if we explicitly pass the
         # bin edges.
         edges, _ = build_histogram(img,256, minval = 0, maxval = 1)
-        result = adjust_histogram(Matching(),img,imgo,edges)
+        result = adjust_histogram(img, Matching(targetimg = imgo, edges = edges))
 
         edges_target, counts_target = build_histogram(imgo, 256,minval = 0, maxval = 1)
         edges_result, counts_result = build_histogram(result, 256, minval = 0, maxval = 1)
@@ -30,7 +30,7 @@
     for T in (Gray{N0f8}, Gray{N0f16}, Gray{Float16}, Gray{Float32})
         img = T.([i < 64 ? i/255 : 0 for i = 0:255])
         imgo = T.([i > 64 && i < 128 ? i/255 : 0 for i = 0:255])
-        result = adjust_histogram(Matching(),img,imgo,256)
+        result = adjust_histogram(img, Matching(targetimg = imgo, nbins = 256))
         edges_source, counts_source = build_histogram(img, 256, minval = 0, maxval = 1)
         edges_target, counts_target = build_histogram(imgo,256, minval = 0, maxval = 1)
         edges_result, counts_result = build_histogram(result,256, minval = 0, maxval = 1)
@@ -42,7 +42,7 @@
         img =  T.(imgg,imgg,imgg)
         imggo = Gray{N0f8}.([i > 64 && i < 128 ? i/255 : 0 for i = 0:255])
         imgo = T.(imggo,imggo,imggo)
-        result = adjust_histogram(Matching(), img, imgo, 256)
+        result = adjust_histogram(img, Matching(targetimg = imgo, nbins = 256))
         edges_source, counts_source = build_histogram(img, 256, minval = 0, maxval = 1)
         edges_target, counts_target = build_histogram(imgo, 256, minval = 0, maxval = 1)
         edges_result, counts_result = build_histogram(result, 256, minval = 0, maxval = 1)
@@ -51,7 +51,7 @@
         # Check that we get the same result if we explicitly pass the
         # bin edges.
         edges = 0.0:0.0019454658031463623:0.4960937798023224
-        result = adjust_histogram(Matching(),img,imgo,edges)
+        result = adjust_histogram(img, Matching(targetimg = imgo, edges = edges))
         edges_source, counts_source = build_histogram(img, 256, minval = 0, maxval = 1)
         edges_target, counts_target = build_histogram(imgo, 256, minval = 0, maxval = 1)
         edges_result, counts_result = build_histogram(result, 256, minval = 0, maxval = 1)
@@ -62,14 +62,14 @@
     # https://github.com/JuliaImages/Images.jl/pull/752
     img = Gray.([0.4 0.45; 0.5 0.55])
     imgo = Gray.([0.1 0.3; 0.7 0.9])
-    result = adjust_histogram(Matching(),img, imgo)
+    result = adjust_histogram(img, Matching(targetimg = imgo))
     edges_target, counts_target = build_histogram(imgo, 256, minval = 0, maxval = 1)
     edges_result, counts_result = build_histogram(result, 256, minval = 0, maxval = 1)
     @test all(counts_target .== counts_result)
 
     # Verify idempotency.
     img = collect(0:1:255)
-    result = adjust_histogram(Matching(),img,img,256)
+    result = adjust_histogram(img, Matching(targetimg = img, nbins = 256))
     edges_target, counts_target = build_histogram(img, 256, minval = 0, maxval = 255)
     edges_result, counts_result = build_histogram(result, 256, minval = 0, maxval = 255)
     @test all(counts_target .== counts_result)
@@ -77,7 +77,7 @@
     # Verify that the algorithm works on integers.
     img = ([i <= 64 ? i : 0 for i = 0:255])
     imgo = ([i > 255-64 && i <= 255 ? i : 0 for i = 0:255])
-    result = adjust_histogram(Matching(),img,imgo,256)
+    result = adjust_histogram(img, Matching(targetimg = imgo, nbins = 256))
     edges_source, counts_source = build_histogram(img, 256, minval = 0, maxval = 255)
     edges_target, counts_target = build_histogram(imgo, 256, minval = 0, maxval = 255)
     edges_result, counts_result = build_histogram(result, 256, minval = 0, maxval = 255)
@@ -86,6 +86,6 @@
     # Verify that the algorithm can cope with NaN.
     img_nan =  Float64.(collect(0:1/255:1))
     img_nan[1:128] .= NaN
-    reference_nan = adjust_histogram(Matching(), img_nan, img_nan)
+    reference_nan = adjust_histogram(img_nan, Matching(targetimg = img_nan))
     @test all(isapprox.(img_nan[129:end], reference_nan[129:end]; atol = 1e-1))
 end
